Add two unit tests to get a feel for things
Just some trivial tests.

I'm wondering if I'm handling the failure case correctly.  That is, if
ParseConfig fails, it returns the zero-value for a Config struct.  But I
am keeping this "IsParsed" attribute on the struct to check whether the
parse was successful.  This doesn't feel write.  It seems like I should
just be checking if the Config object is the zero-value, but Go says I
can't compare struct objects that have strings... Hmmmm....