Yeah! SSH library proof-of-concept working
The geto SSH library can now connect to a remote host.  No password is
necessary, as the library uses public key authentication.

The geto.ini has been updated to include the private key file, which the
geto SSH library uses to get the public key.

The code is currently hard-coded with a username and URI to access a VM
at my desk, this should be fixed, obviously, to use dynamic information.