Add ToFile() method to Task
The ToFile() method creates a file in the local work path.  The filename
format is <task id>_<task script name>.  This file will be how the task
is executed on the remote host, and the file format will allow for finding
remote processes based on unique task ID or by common script names.