Begin work on remote locking mechanism
The Script struct includes a maxConcurrent value that is meant to limit the
number of the same scripts running on a given target.  This changeset
contains the beginnings of implementing the ability to limitt he number
of concurrent scripts by creating a remote lock.

This changeset also introduces the idea of a wrapper task running the
actual user task.  The user task is executed by the wrapper task, and
the user task's output is put in special "stderr" and "stdout" files in
the task directory.